# backend/openapi.yaml
openapi: 3.0.3
info:
  title: VitalTrack API
  version: 0.1.0
  description: API para gestión de usuarios, perfiles, dispositivos y actividades.
servers:
  - url: http://localhost:3000
...
paths:
...
  /api/actividades:
    get:
      tags: [Actividades]
      summary: Listar actividades
      parameters:
        - in: query
          name: usuario_id
          schema: { type: integer }
      responses:
        '200': { description: Lista, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Actividad' } } } } }
    post:
      tags: [Actividades]
      summary: Crear actividad
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ActividadCreate' }
      responses:
        '201': { description: Actividad creada, content: { application/json: { schema: { $ref: '#/components/schemas/Actividad' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/actividades/{id}:
    get:
      tags: [Actividades]
      summary: Obtener actividad
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Actividad, content: { application/json: { schema: { $ref: '#/components/schemas/Actividad' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Actividades]
      summary: Actualizar actividad
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ActividadUpdate' }
      responses:
        '200': { description: Actividad actualizada, content: { application/json: { schema: { $ref: '#/components/schemas/Actividad' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Actividades]
      summary: Eliminar actividad
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '204': { description: Eliminado }
        '404': { $ref: '#/components/responses/NotFound' }
components:
  responses:
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    BadRequest:
      description: Datos inválidos
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: object }
...
    Actividad:
      type: object
      properties:
        id: { type: integer }
        usuario_id: { type: integer }
        tipo: { type: string }
        hora_inicio: { type: string, format: date-time }
        hora_fin: { type: string, format: date-time }
        duracion_segundos: { type: integer }
    ActividadCreate:
      type: object
      required: [usuario_id]
      properties:
        usuario_id: { type: integer }
        tipo: { type: string }
        hora_inicio: { type: string, format: date-time }
        hora_fin: { type: string, format: date-time }
        duracion_segundos: { type: integer }
    ActividadUpdate:
      allOf:
        - $ref: '#/components/schemas/ActividadCreate'